requirements.txt
boto3
pandas
numpy
streamlit
spotipy
seaborn
plotly
scikit-learn
python-dotenv
notebook
transformers
torch
PyYAML
duckdb
